Updates to Launcher app launching
It now launches when the object is tapped. It does not rotate and does not disappear when you get close to it.

You can now change name of an object by using the hold gesture on it to open up the system keyboard. Requires the project to be deployed as XAML instead of D3D (not sure what tradeoffs that entails)