Base explodeview going on
Usage: Attatch the ExplodeObejct script to an empty gameobject and make sure the rigidbody is kinematic, then when you tap it each of its children will explode away from the center of the original object. Works with layers, like an onion.

Should make it into an interface, so the Explode Object can be implemented separately